% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{s7contract}
\alias{s7contract-package}
\alias{s7contract}
\title{s7contract: Contract Helpers for S7}
\description{
\code{s7contract} provides experimental runtime contracts and generative laws:
}
\details{
\itemize{
\item Go-like structural interfaces defined by required generics.
\item Rust-like explicit traits with default methods and associated metadata.
\item Property-based laws with integrated shrinking and tinytest expectations.
}

The package keeps actual method dispatch inside ordinary S7 generics and uses
runtime checks to describe or assert conformance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sounkou-bioinfo/s7contract}
  \item \url{https://sounkou-bioinfo.github.io/s7contract/}
  \item Report bugs at \url{https://github.com/sounkou-bioinfo/s7contract/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

}
